engenhariaSoftware/src/main/java/org/example/AlunoManagerTest.java
package org.example;

import static org.junit.jupiter.api.Assertions.*;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;


public class AlunoManagerTest {

    private AlunoManager alunoManager;

    @BeforeEach
    public void setUp() {
        alunoManager = new AlunoManager();
        alunoManager.cadastrarAluno("João", "Matemática", 8.5);
        alunoManager.cadastrarAluno("Maria", "Português", 9.0);
        alunoManager.cadastrarAluno("José", "Física", 7.0);
    }

    @Test
    public void testBuscarPorNomeExistente() {
        Aluno aluno = alunoManager.buscarPorNome("Maria");
        assertNotNull(aluno);
        assertEquals("Maria", aluno.getNome());
    }

    @Test
    public void testBuscarPorNomeInexistente() {
        Aluno aluno = alunoManager.buscarPorNome("Carlos");
        assertNull(aluno);
    }

    @Test
    public void testBuscarPorMateriaExistente() {
        alunoManager.cadastrarAluno("Ana", "Matemática", 6.5);
        alunoManager.cadastrarAluno("Pedro", "Matemática", 7.5);

        alunoManager.buscarPorMateria("Matemática");
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testBuscarPorMateriaInexistente() {
        alunoManager.buscarPorMateria("Química");
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testBuscarPorNotaExistente() {
        alunoManager.cadastrarAluno("Carlos", "História", 7.5);
        alunoManager.buscarPorNota(7.5);
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testBuscarPorNotaInexistente() {
        alunoManager.buscarPorNota(10.0);
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testListarAlunosPorNota() {
        alunoManager.cadastrarAluno("Ana", "Química", 7.0);
        alunoManager.cadastrarAluno("Pedro", "Matemática", 7.5);

        alunoManager.listarAlunosPorNota(7.5);
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testListarAlunosPorMateria() {
        alunoManager.listarAlunosPorMateria("Física");
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }

    @Test
    public void testListarMateriasDeAluno() {
        alunoManager.cadastrarAluno("Ana", "Matemática", 6.5);
        alunoManager.cadastrarAluno("Pedro", "Matemática", 7.5);

        alunoManager.listarMateriasDeAluno("Pedro");
        // Neste caso, verificamos a saída no console manualmente
        // Como este é um teste de saída no console, não podemos automatizar a comparação aqui diretamente
    }
}